% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uid_generate.R
\name{uid_generate}
\alias{uid_generate}
\title{Generate random ID}
\usage{
uid_generate(char_num = 6)
}
\arguments{
\item{char_num}{Number of characters in a random ID}
}
\value{
a character vector
}
\description{
Generate random ID
}
\examples{
uid_generate(char_num = 10)
}
